* [pre-commit.ci] pre-commit autoupdate (#2107)


* Update publication list with new blog (#2111)

* Update publication list with new blog


* Update publication list num


---------


* update the License (#2108)


* Add autotune support for PT2E (#2110)

Add autotune support for PT2E and disable some conv1d-related test on HPU
---------


* Add intel-extension-for-pytorch to Transformers-like API requirements (#2113)


* Add VLM quantization & loading into transformers-like API (#2116)


* Fix hf_device_map setting for transformers-like api (#2122)


* add mapping entry to 1.20 (#2126)


* fix bug of lwq gtpq (#2128)


* add xfail for onnx test_layer_wise.py (#2129)

* Doc: Update fp8 accuracy test data and update docker image 1.20.0 (#2130)


* [SW-219274] - Changing the quant method name in lm-head (#150) (#2132)

* [SW-219274] - Changing the quant method name in lm-head (#150)

* Update helper_modules.py

---------


* Adapt ipex xpu transformers version (#2134)


* Add phi3 vlm  transformers example (#2135)


* fix saving issue for group_size=-1 (#2138)


* Add v3.3 release faq (#2139)


* bump release version into 3.3 (#2140)


* upgrade numpy from 1.23.5 to 1.26.4 (#2115)


* Update publications (#2145)

* update publications


* Add transformers to align onnxruntime-extensions=1.14.0 (#2147)


* Freeze 2x package versions (#2151)


* fix vulnerability (#2149)


* Bump into v3.3.1 (#2152)


* [SW-218939] fix memory mapping failure in UT (#2154)


* [SW-223106] change code with robust implementation (#2153)

* fix ILITV-3854


* fix sdxl_smooth_quant


* workaround for ILITV-3858


* fix ILITV-3859


* add xfail for torchvision


* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

+++ examples/3.x_api/pytorch/recommendation/dlrm/static_quant/ipex/dlrm_s_pytorch.py
@@ -845,7 +845,7 @@ def run():
         example_inputs = (X_test, lS_o_test, lS_i_test)
         assert args.inference_only, "Please set inference_only in arguments"
         from neural_compressor.torch.quantization import StaticQuantConfig, autotune, TuningConfig
-        tune_config = TuningConfig(config_set=StaticQuantConfig.get_config_set_for_tuning())
+        tune_config = TuningConfig(config_set=StaticQuantConfig.get_config_set_for_tuning(dtype="int8"))
 
         dlrm = autotune(
             dlrm,
